2004-11-20  Al Riddoch  <alriddoch@zepler.org>
	* client/CreatorClient.cpp, rulesets/Py_BBox.h, rulesets/Py_Optime.h,
	  rulesets/Py_Point3D.h, rulesets/Py_Quaternion.h,
	  rulesets/Py_Vector3D.h: Re-order Python.h includes to get rid of
	  some stupid warnings.

	* rulesets/Py_World.cpp: Add a cast when passing float time to
	  WorldTime which takes in int.

+++ rulesets/Py_World.cpp
@@ -24,7 +24,7 @@ static PyObject * World_get_time(PyWorld *self)
     if (wtime == NULL) {
         return NULL;
     }
-    wtime->time = new WorldTime(self->world->getTime());
+    wtime->time = new WorldTime((int)self->world->getTime());
     wtime->own = true;
     return (PyObject *)wtime;
 }
